Fix spaceid_to_slice()
The main problem was that the slice should not include the stride from the
spaceid if the block is not 1, because blocks are not strided. We require one
of block or count to be 1, so this isn't a problem (as far as I can tell,
h5py has this restriction as well).