Start fixing creation of datasets
Prior to this, datasets are created in the group of the previous version. But
this actually corrupts the previous version if any new datasets are created.
Instead, we should copy everything over from the previous version. It also
means we shouldn't actually call create_dataset() on the h5py Group object,
but rather fake it so that it goes through our machinery.

This isn't completely fixed yet. There are still some bugs, and we need to
handle groups properly. There are also some tests that need to be fixed, as
well as new tests that need to be written.