Fix reading from hdf5 files that have been moved
h5py by default hard-codes the filename as the virtual data source, meaning if
the file is renamed, it won't be able to find the virtual source, and will
give all 0s (because it masks the error from HDF5). If we instead pass the
undocumented string "." as the file name, it references the current file and
works after renaming.

Fixes #28.