Initial support for sparse datasets
Sparse datasets have chunks that are not stored directly in the file. The data
in those chunks is assumed to be equal to the fill value.

This supports writing sparse datasets equal to the fill value
(create_dataset() with data=None), but it does not yet support writing partial
sparse data. Reading values is also not yet supported except when the file is
closed and reopened. Support for these features requires some features to be
added to ndindex first.

This is based on #62, which can be merged independently.

See issue #60.