Make create_virtual_dataset() and write_dataset() use a dictionary of slices
Instead of a list of slices which implicitly used the index as the chunk, it
is now a dict that maps ndindex multidimensional chunk indices to a slice of
the raw data.

I haven't yet updated write_dataset_chunks as this will also require updating
the data_dict in the wrappers to use this as well.