Implement reading for InMemoryDataset
We have to switch to using InMemoryDataset.__getitem__ instead of
InMemoryDatasetID.read. The read() method gets h5py spaceid objects that do
not contain enough information to do the read. In particular, dimensions that
are removed in the sliced array (integer indices) are removed from the array
that we read to. But they are represented as a length-1 slice in the fspace
selection. There is no way to distinguish between an integer index and a
length-1 slice in the fslice selection. Therefore, properly aligning the
selection onto the input array is not possible (or if it is possible it is
very difficult), because we do not know which length-1 dimensions in the
fspace were removed from the in-memory array.

Performing the operation on InMemoryDataset.__getitem__ is much simpler now
that we have ndindex because we can operate on the input index directly,
rather than trying to translate back from HDF5 selections, which have
different semantics than NumPy indices.

This still needs tests, and cleanup to the InMemoryDatasetID code to remove
the old read() method.