Use a single file for all the test combinations in test_resize_multiple_dimensions()
This makes the test run in about 2 minutes, rather than 14 minutes.
Unfortunately, it also causes the resulting temporary test file to be about 1
GB in size by the time the test finishes, but I consider this to be an
acceptable tradeoff since the speedup is so drastic. I don't want to reduce
the test itself since it has caught a lot of bugs already.