Don't store data in the virtual layout if the slice is empty
This causes later calls to get_virtual_sources() to error. If there are no
slices, this will mean no virtual sources will be set, and
create_virtual_dataset() will actually create a nonvirtual dataset. The places
that assume that a dataset is virtual have been updated to handle this case.

This also adds the raw_data name as an attribute to versioned datasets, since
it cannot be extracted from the HDF5 metadata when it is not actually virtual.