Handle sparse data in InMemoryDataset
This makes sparse data work for InMemoryDataset (e.g., for reading a sparse
dataset from a file rather than from memory), but it still only works for the
trivial case (no non-fillvalue) in for InMemoryArrayDataset, i.e.,
InMemorySparseDataset.

That is to say, sparse data (data=None) works, but only if you close and
reopen the file before writing any non-fillvalue data to the dataset.

This also assumes that the sparsity of the dataset is not too high, as it
iterates every chunk of the data. If the sparsity is high, then every
read/write operation will be slow (see the discussion on #63).