Fix as_subchunks only iterating the chunk index
This is a change that is made in Quansight-Labs/ndindex#110, and so this
commit will only work against that branch. The change was made to improve the
performance in many cases where the extra index that was yielded from
as_subchunks() was not needed and therefore did not need to be computed.