Comment out an assertion in create_virtual_dataset
The assertion is a good sanity check, but it can be slow in general, as
newshape() is can be slow to compute currently (it was taking over a third of
the total time of the function). Also, the correctness of the assertion should
generally be assured by the correctness of as_subindex, so it is not as
important to have now that we are more confident in that function than when it
was originally written.