Correctly handle virtual datasets in swap()
We need to recreate them. If we just move them, they will still point to their
original raw_data dataset. Fortunately, this can be done more efficiently than
creating the virtual dataset in the first place, as we only have to copy the
existing data and modify the source dataset name.