Fix the str dtype not being maintained from the fillvalue
This used to work because it used to always add trivial chunks, which ended up
maintaining the dtype, but now that it is more efficient, it doesn't do that
any more. The correct behavior is to maintain the dtype, since self.fillvalue
loses that information, basically because of this:

>>> np.array([b''], dtype='S5')[0].dtype
dtype('S')

Fixes #161.