Perform the virtual layout selection manually
The h5py code does a deepcopy of the VirtualSource object whenever it is
indexed. But this is slow, and only done so that the it can produce a new
object to pass the data through to the VirtualLayout in __setitem__. We just
direcly run the code from VirtualSource.__getitem__ and
VirtualLayout.__setitem__, sans the deepcopy(), which creates HDF5 selection
objects and stores them on the VirtualLayout.