Create a wrapper dataset object
This allows converting an InMemoryDataset into an ArrayDataset inplace. This
will be more efficient when most of the chunks are already read into memory.
Presently this is only done in the case when the entire dataset is set with a
single index, like dataset[:] = ..., but this can easily be extended in the
future to other cases when a majority of the chunks are already read into
memory.