Use len() instead of as_subindex() to compute the chunk size for the virtual dataset
We know the chunk index is always a "chunk index", meaning it is a tuple of
slices with positive start and stop and step 1. So as_subindex is overkill.
The len() is easier to read and, more importantly, faster. We could also use
stop - start, but that wouldn't be much faster, and would be less obviously
correct.