Cache part of the computation in spaceid_to_slice
If one dimension is small relative to the chunk size, then most slices for it
will be the same. Since the data_dict has indices for chunk in the dataset,
this avoids duplicate Slice construction.

The entire tuple is not cached because it will generally be unique. Only in
the case where the underlying data is copied will it be reused, but this
typically doesn't happen within the same dataset except in toy examples.