Workaround the upstream h5py garbage collection issue
This is the issue from https://github.com/deshaw/versioned-hdf5/issues/162.
h5py has a fix, but we won't be able to use it in 2.10.0 unless we manually
backport the patch and custom build h5py. The bug doesn't affect the actual
tests, it just causes them to crash sometimes when the file is closed.