Read directly from the virtual dataset when there are no arrays
This happens in particular when reading from an already committed version,
since everything will be read-only and already written to the file. This is
much faster than the previous method because the data dict does not need to be
constructed, and we do not have to read in each chunk one at a time in a
Python loop.