Fix some issues with h5py 3.3
The sources attribute for VirtualLayout was removed because VirtualLayout now
creates the virtual sources property list directly, so some code that was
manipulating that needed to be rewritten. The upside is that this new
VirtualLayout code should use less memory in cases where a dataset with many
chunks is created.

See https://github.com/h5py/h5py/pull/1905.