Add new version of delete_versions()
This should be faster than the old version, because it deletes the chunks more
directly rather than trying to recreate all the datasets.

However, there are still some performance issues with this compared to the old
version.

This also renames "delete_version()" to "delete_versions()", but keeps the old
name for backwards compatibility.