Don't return wrapper classes when the file is in read-only mode
This improves the read performance in these cases. When the file is not in
read-only mode, we still use wrapper classes even for committed groups because
these have the safety checks to prevent accidental writing, which would
corrupt the raw data.

This breaks using mask indices in h5py 2.10 for read-only files, since these
are not supported for virtual datasets and we only support them manually via
the wrappers. h5py 3 does support them, however.