Improve the performance of create_virtual_dataset for h5py 3
Using VirtualLayout and VirtualSource directly is slow for a few reasons. The
main reason is that VirtualSource does a deepcopy of itself every time you
slice it. The high-level code also calls the internal helper
_convert_space_for_key(), which is unnecessary for our given slices (it only
does anything for unbounded slices), and recreates a selection object each
time, which we can do just once.

Any further performance improvements here would have to come from speeding up
the h5py select() function, which is where most of the time is now spent.