Faster and correct chunk reuse logging (#324)
Closes #323.

The logic in `_log_version_diff_stats` is currently not correct
since it just compares the new `hashtable` with itself. This
ends up always logging that all chunks were reused, which is obviously
not correct. Furthermore, the size of the hashtable also does not
necessarily correlate with the number of chunks reused for a
dataset version.

It is also fairly slow since loading the hashtable is expensive
if it drops out of the lru_cache (and unnecessary anyway, as discussed
above). The calculations also happen regardless of the logging level,
which means that they are mostly wasted anyway.

This change simplifies the logging of version diff stats by doing
it directly when the dataset is written and when we can observe
the number of added reused/chunks.